requirements.txt
pymongo==4.3.3
requests==2.26.0
pyjwt==2.8.0